+++ chrome-cordova/chrome-apps-api-tests/www/background.js
@@ -1,0 +1,5 @@
+chrome.app.runtime.onLaunched.addListener(function() {
+  console.log('onLaunched fired.');
+  var helper = cordova.require('cordova-plugin-chrome-apps-test-framework.app_helpers');
+  helper.createUiWindow();
+});