+++ chrome-cordova/chrome-apps-api-tests/www/main.js
@@ -1,0 +1,14 @@
+(function() {
+
+'use strict';
+
+/******************************************************************************/
+
+document.addEventListener("DOMContentLoaded", function() {
+  var cdvtests = cordova.require('cordova-plugin-test-framework.main');
+  cdvtests.init();
+});
+
+/******************************************************************************/
+
+}());